Win64 now uses only 4 registers to transfer arguments

+++ mcsema/BC/Util.h
@@ -45,7 +45,7 @@ extern llvm::IntegerType *gWordType;
 extern llvm::Module *gModule;
 
 
-llvm::Value *GetConstantInt(uint64_t size, uint64_t value);
+llvm::Value *GetConstantInt(unsigned size, uint64_t value);
 
 // Return the type of a lifted function.
 llvm::FunctionType *LiftedFunctionType(void);
